Use hash-based IDs for offer lookups
Switch from random UUIDs to deterministic hash-based IDs derived from URLs. This enables fast point reads for duplicate detection.

- Generate offer IDs from URL hash instead of random UUIDs
- offer_url_exists() uses point read instead of query
- Duplicate detection moved to spider parse() methods
- Removed SkipDuplicates from active pipelines
